+++ src/pytest_bdd/compatibility/allure.py
@@ -1,5 +1,5 @@
-import importlib.util
+# ALLURE_INSTALLED = all(
+#     (importlib.util.find_spec("allure_commons") is not None, importlib.util.find_spec("allure_pytest") is not None)
+# )
 
-ALLURE_INSTALLED = all(
-    (importlib.util.find_spec("allure_commons") is not None, importlib.util.find_spec("allure_pytest") is not None)
-)
+ALLURE_INSTALLED = False